stockWeb/static/stockWeb copy/requirements.txt
Flask
yfinance
pandas
numpy
scikit-learn
ta
nltk
vaderSentiment
python-dotenv
joblib